+++ controllers/userController.js
@@ -1,0 +1,31 @@
+// Route Handlers
+exports.getAllUsers = (req, res) => {
+  res.status(500).json({
+    staus: 'error',
+    message: 'This route is not yet defined!',
+  });
+};
+exports.getUser = (req, res) => {
+  res.status(500).json({
+    staus: 'error',
+    message: 'This route is not yet defined!',
+  });
+};
+exports.createUser = (req, res) => {
+  res.status(500).json({
+    staus: 'error',
+    message: 'This route is not yet defined!',
+  });
+};
+exports.updateUser = (req, res) => {
+  res.status(500).json({
+    staus: 'error',
+    message: 'This route is not yet defined!',
+  });
+};
+exports.deleteUser = (req, res) => {
+  res.status(500).json({
+    staus: 'error',
+    message: 'This route is not yet defined!',
+  });
+};
